cman: Don't sent a RECONFIGURE notification until we have a valid config
Before this patch, cman would send a RECONFIGURE notification as soon
as it got a message that a new config file should be read. If that
failed then applications would get a new RECONFIGURE message every second
until it succeeded.

Now they just get one, at the end of the process.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>