Non-default block size confuses gfs2_grow
bz 469773 (originally)

When gfs2_grow calculates the number of new resource groups
needed for an extended file system, it was doing so using values
calculated before the superblock was read in.  Therefore it used
the default block size (4K) regardless of the actual block size.
This patch moves the call to fix_device_geometry after the
superblock has been read in so the calculations are done using
the correct block size.