libgfs2: Forgot to remove headers from apps
The applications were including some of the headers directly,
and since they didn't (for some reason) rebuild when I removed
the headers, I missed that fact.

In fact one of the utils had included the content of ondisk.h
four times over into a single c file. I think we can just rely
of libgfs2.h for that now.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>