[FENCE] Fixes ##485065 - Add support for LPAR/HMC v3
New options for LPAR agent: (-H; hmc-version) and (-c; command-prompt). HMC version
have to be set because v3 and v4 are not compatibile. It is possible to auto-detect
version (lshmc -V) but this is quite slow to do it in each fencing operation. Default
value is '4' because we don't want to change previous behaviour. It does not mean
that it should not work with HMC v5, v6 - we just need to mark them.

Command prompt option was set to default value but missing in device's options. Older
LPAR has different prompt so it was added to default one. If you wish to add your own
prompt that remember that $ has to be written as '\$'.