cman: Don't eject nodes that were quorate on their own if we have no state
If a node boots up and becomes quorate on its own (eg because of
expected_votes=1 or qdisk) and another node then joined in the cluster, the
new node would kill the first node because it detects that it had state and
was not previously part of the cluster.

Now it checks to see if the new node also has state. If it does not then
the cluster join is allowed because we are joining a stateful node to
a stateless node and that should be allowed.

bz#485026

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>