libfence/fence_node: unfencing
The new libfence unfence_node() function looks for per node <unfence>
sections.  The <unfence> section contains no methods, only devices.
unfence_node() attempts all devices in the unfence section, regardless
of whether any fail.  The function returns success only if all devices
were successful.

The <device> line within <unfence> would typically require action="on"
or similar, to tell the agent to do the reverse of normal fencing.
The arg string for agents are constructed in the same was as for fencing:
the combination of the per-node unfence/device args and the fencedevice
args.

The fence_node command has a new -U option that causes it to call
unfence_node() rather than fence_node().

Signed-off-by: David Teigland <teigland@redhat.com>