fenced: check for group/groupd_compat at startup
It was being read along with the other ccs values, which happens at
domain join time, not at startup like other daemons.  groupd_compat
needs to be checked at startup.

Also remove the check it was doing that the local cman node name is
in ccs, I don't think that's ever possible.

Also fix the section name to <group/> in groupd man page.

Signed-off-by: David Teigland <teigland@redhat.com>