init: stop daemons before cman
commit b263da25536c90e6ca654a0be6e64481fe31eba2 did the right thing
at that time, but it should have been updated in time with the evolution
of the rest of the stack.

At stop/shutdown time, fence_tool/fenced will now refuse to stop if
there is an active mount and therefor the init script will not continue
as it used to do before.

As a consequence, with a clean shutdown operation, all daemons will now
release the resources and not leak ipc resources.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>