libgfs2: Remove 'die' calls from __get_sysfs
This patch removes the calls to die from __get_sysfs and adds error
handling to it and its callers. To enable this, the prototype of
get_sysfs_uint had to be changed so that error status can be returned.
Callers to get_sysfs and get_sysfs_uint are also updated to handle error
cases where appropriate.

Signed-off-by: Andrew Price <andy@andrewprice.me.uk>