gfs2_convert results in GFS2 File System Corruption
bz 490136

The gfs2_convert tool was taking an unreasonable amount of
time and memory, and if your files were big, the could be
corrupted as well.  This patch reworks the functions that
convert dinode data to rearrange the indirect block pointer
data so that it does it quickly, correctly and with minimum
memory usage.