cman init: do a better check to umount configfs
dlm is a configfs user. Once we modprobe dlm and mount configfs,
the "ls -1" check will always avoid configfs umount.

Do a very soft attempt to remove the kernel modules we load without failing
if they are in use. If we succeed to remove the modules and there are no
more users for configfs, then umount it.

One catch 22 note is that we need to modprobe -r configfs only after it is
umounted, so perform that operation exclusively after we umount.
No failure is reported or generated as there might be other users around.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>