Fix block count in pass1b.
bz 493727 -  GFS: gfs_fsck can delete everything in a corrupt file system

When duplicate blocks are discovered by gfs_fsck, the block count
was not being saved by the duplicate processing code.  That caused
the fsck to think there were zero blocks in the directory.  The
descrepency caused the whole root directory to be deleted.