group_tool: -g option for controlling query
Allow the same -g0, -g1, -g2 style option that all daemons use to
select the old/new mode for queries.  0 for new, noncompat queries
to individual daemons, 1 for old compat query to groupd, 2 to
dynamically use 0 if daemons running in new mode or 1 if groupd
is running in old mode.  Default is 0.

Signed-off-by: David Teigland <teigland@redhat.com>