cman init: faster stop operation
stop daemon operation was waiting for one sec in between
each check, before going forward. This was a bit too much.

On distributions that have a sleep that supports decimal
operations, switch to 250ms instead of 1 second.

This makes the stop operation a lot faster (in the normal
case).

Tested by David Teigland.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>