rgmanager: Make vm.sh use libvirt
This makes rgmanager use virsh/libvirt instead of xm so
it can manage either KVM or Xen virtual machines.

It will still use xm if a user has a non-standard config
file path, however, the recommendation is that the
user store all Xen configuration files in one place and
use bind mounts to mount the requisite path on /etc/xen.
Virsh does not have a notion of a search path like xm,
therefore, further support of the path attribute in
xm will be limited.

During this partial overhaul, extraneous attribute
parsing was removed and the 'stop' phase now correctly
returns failure if it cannot contact the hypervisor.

Users may force the use of 'xm' by adding:

  use_virsh="0"

to vm resources.  Furthermore, users may force use
of a specific hypervisor by using:

  hypervisor="qemu"

or:

  hypervisor="xen"

The other important note is that libvirtd is now
required to be running in order to operate with the
virsh command set even if using the Xen hypervisor.

Resolves: 412911 468691

Signed-off-by: Lon Hohberger <lhh@redhat.com>