fence: Allow IP addresses as node names
When checking if a node is a member of the cluster, fenced does a full-string
check of the name. If that doesn't match it then assumes that one of the
names is a FQDN and one is truncated so it starts checking the bit before the
first dot.

If the node names are IP addresses then this will match all of the time for
most clusters. eg: 192.168.2.1 will match 192.168.2.2 because it just
matches the 192. This makes fenced think that the node has rejoined the cluster
because it sees the name in the list of active nodes, and won't fence it.

This patch abandons the match check afer the full-string one if the node
name is found to be an IP address.

bz#504158

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>