mkfs.gfs2: Remove dep on libvolume_id
We had been using libvolume_id to look at the content of a
device or file before mkfs makes any changes to give the
user a hint that they are really overwriting the correct
thing.

Since libvolume_id has been removed from rawhide, Fabio
has asked that we remove the dependency from here.

In order to do that, I've replaced the call to libvolume_id
with a fork()/exec() of file(1) from the GNU file utils. This
can identify a wider range of files/fs formats and thus will
print a more informative message. There is no new dependency
created though: if file(1) doesn't exist then we just print
a message saying that we can't identify what is in the device/file
and carry on as normal.

This might be a bit rough at the edges, but it does the job
and should fix the problem for now.

Signed-off-by: Steven Whitehouse <swhiteho@redhat.com>