GFS2: fsck.gfs2 sometimes needs to be run twice
bz 500483

This patch fixes a lot of fsck.gfs2 bugs.  Most of these bugs
have to do with improper bitmap setting that resulted in needing
to run fsck.gfs2 a second time in order to fix the file system.
There were also cases where changes were not written out and
cases where the user was not asked for permission to make changes.
There were also fixes related to extended attributes.