fence agents: Option for setting port for telnet/ssh/ssl used by fence agent
On some fencing devices there is possibility to change default port where
ssh/ssl/telnet/... is listening. There is new option 'ipport' (getopt: -u)
which can be used for this purpose. It is probable that it will make option
'udpport' (also -u; currently used in SNMP based agents) obsoleted in one
of the next versions.

Resolves Red Hat Bugzilla #496446