cman: make cman_tool leave remove work even when no services are running.
cman_tool leave takes a shortcut internally if no services are listening
for shutdown events. Unfortunately, a bug in this shortcut means that the
leave flags gets lost, so quorum is not adjusted downwards.

This patch fixes that shortcut. I should make it clear that 'leave remove'
works fine if any cluster services are running.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>