cman: Fix a situation where cman could kill the wrong nodes
hmm, how to describe this .... Hmmmm. OK lets try:

There were a couple of places in the cman code where the transition message
assumed that the node in question (either this node or the sending node) was
joining the cluster, rather than just sending it's current post-transition
state. This was wrong. It's a common problem we have with openais/corosync
in that it always merges clusters rather than joining from scratch so we
need to detect that in some way.

The code in ais.c has a flag called 'first_trans' which it sets when it first
encounters another node in the cluster. We should use this more often as it's
really helpful. So this is what we now do. The comments in the existing code
make it clear that it assumed the node was joining and not just part of an
existing transition, but the first_trans flag was not checked, so it was
fairly obvious what was going on.

So, now we check the first_trans flag in all places where the code assumes
that the node is joining a new cluster.

See ?

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>