cman init: fix cman startup issues
with recent changes to cman/corosync to better report errors
on startup, we accidentally fixed "too much" logging with debug="on".

those changes introduced a bunch of problems with file descriptors
in the init script.

change the init script to startup cman in such a way that it
won't make a mess of fd's.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>