cman init: integrate ocfs2_controld.cman startup
with the most recent ocfs2-tools, it is possible to mount
ocfs2 using a cman cluster.

This change address the scenario where o2cb starts before cman.

In that specific case ocfs2_controld.cman will fail
to start from o2cb because cman is not yet available.

Check if o2cb is running and configured as cman cluster,
if so attempt to start ocfs2_controld.cman.

Signed-off-by: Fabio M. Di Nitto <fdinitto@redhat.com>