dlm_controld: fix start matching for partition+merge changes
When a node is removed, added, removed due to a partition+merge,
the start messages for the second removal are mistakenly matched
to the first removal (since the change descriptions are idential).
To prevent this, detect when there are identical outstanding
changes and send a start+nack for the first before sending the
regular start for the second.

Signed-off-by: David Teigland <teigland@redhat.com>