dlm_controld: change default enable_quorum to 0
Fencing waits for quorum when necessary, and dlm_controld waits
for fencing.  So, there's no reason that dlm_controld needs to
wait for quorum itself, apart from the possible expectation of
users/apps that a lockspace (and cluster fs) will not be enabled
in an inquorate cluster.

Not waiting for quorum does allow a lockspace (and cluster fs)
to be fully enabled/started/recovered in an inquorate cluster as
long as there is no fencing outstanding.  The big advantage
is that broader shutdown is possible in an inquorate cluster,
i.e. lockspaces (and cluster fs's) can be left (unmounted) in
an inquorate cluster.

Shutting down a node involves the others disabling/enabling the
lockspace, and the enabling step is what will block on quorum,
which then prevents the shutdown on those nodes.

Signed-off-by: David Teigland <teigland@redhat.com>