cman: Add a validation step to cman_tool join & version subcommands
cman_tool join & cman_tool version now run the configuration through
the schema validator before continuing. By default they will abandon
the operation if they encounter an error. This behaviour can be changed
wit the -D switch to both subcommands.

In addition cman_tool version will also call ccs_sync to distribute the
configuration file if it sees that xmlconfig is being used. This can be
prevented by adding the -S switch.

As it stands, this only really works with xmlconfig because the other
supported configuration (LDAP) needs additional environment variables
to work and these are currently not stored in the corosync objdb. I will
fix this in a future patch.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>