config: fix LDAP loader
The LDAP loader was broken by the change from cn=cluster to name=cluster.

The fix is slightly hacky, it simply bypasses the "name=" check on the first
LDAP entry, which is guaranteed to be "name=cluster". At least, if it isn't
then the whole configuration is badly broken anyway.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>