fenced: fix reported fence time
The fence time reported through libfenced was the fence time on
the master node.  If the node times are different this can break
the dlm_controld check that the fence time is later than the time
it saw the node start.  Instead, save the local time when we get
the victim_done message and report that.

(I've considered this issue in the past but somehow didn't believe
there was a problem... don't recall that reasoning, so not entirely
sure there's not still something being missed.)

Signed-off-by: David Teigland <teigland@redhat.com>