libfence: add nodename to agent args (for unfence)
Add nodename of victim to agent args if it is not set explicitly
as a per-node device arg.

Some agents in the past, and probably the future, will want the
nodename of the victim.  It's so simple to add automatically that
there's no reason not to always include it, and avoid requiring
people to sometimes duplicate the name in the fence section.

(same as previous commit but for unfence args)

Signed-off-by: David Teigland <teigland@redhat.com>