libfence: increase FENCE_AGENT_ARGS_MAX to 4096
The previous limit of 1024 may be a bit small in some cases.

Also fix and clean up the handling of terminating \0 in the
agent and args strings.  The MAX values now include space
for the terminating \0, and snprintfs are responsible for
ensuring a \0.

(Have tested and verified that the fence_log struct size change
does not cause compatibility problems.)

Signed-off-by: David Teigland <teigland@redhat.com>