rgmanager: Use RG_START_RECOVER after relo failure
rgmanager correctly tries to restart a service all
around the cluster in the event of failures when
using RG_ENABLE (e.g. clusvcadm -e), but if a
relocation operation fails, the service can end
up stuck in the 'recovering' state.  This is because
we were not switching to the RG_START_RECOVER
operation after the first relocation attempt failed.

Resolves: rhbz530409

Signed-off-by: Lon Hohberger <lhh@redhat.com>