fsck.gfs2: make query() count errors_found, errors_fixed
This patch changes the query() function so that it keeps track
of errors_found and errors_corrected.  This enforces a consistent
count of errors found and corrected so that the return code from
fsck.gfs2 is correct and future code changes that require calls
to query don't have to code it.  It also makes the calls a tiny
bit faster because they no longer have to pass the opt variable.

rhbz#455300