GFS2: fsck.gfs2 should fix the system statfs file
This patch allows fsck.gfs2 to check the statfs file values
against the real size of the file system by running through
the resource group information.  If found to be incorrect,
it fixes the statfs file.  To facilitate this, the function
do_init in libgfs2 was broken apart into its constituent
pieces to build the inum file and statfs file and mkfs.gfs2
was changed to use the two functions that replace it.

rhbz#539337