Attach buffers to rgrp_list structs
This patch attaches a buffer_head to every rgrp_list in memory.
This greatly improves performance because it no longer needs to
get and put the buffers for the rgrps all the time.  Since the
buffers are linked to the rgrp list, they no longer need to be
linked to each other, so it helps untangle the buffer_head mess.

rhbz#455300