cman: Make cman quit if it has an old config when it starts up
If cman starts up with an old configuration version, it now waits for
a configured time (default 10 seconds) to get a more current one.
If it doesn't get it then it will quit to allow the node to
continue starting up.

This is mainly because at boot-up it's not possible to update the configuration
file before cman starts as ricci is started afterwards, so starting
with an old configuration would cause cman to wait forever looking for
an updated config.

Signed-off-by: Christine Caulfield <ccaulfie@redhat.com>