gfs_fsck: Don't try to convert free_meta to free if -n was specified.
The code was checking free_meta blocks and counting them as errors
even if -n was specified on gfs_fsck.  But free_meta blocks aren't
technically an error and gfs_fsck's converting them to free blocks
is a convenience, so if -n was specified, don't treat them as errors.

rhbz#508978