gfs_fsck: Don't try to convert free inodes to free if -n was specified.
The code was checking free inodes and attempting to convert them
to free blocks, even if -n was specified on gfs_fsck.  But free inodes
aren't technically errors and gfs_fsck's converting them to free blocks
is a convenience, so if -n was specified, don't treat them as errors.

rhbz#508978