fsck.gfs2: Verify rgrps free space against bitmap
This patch makes fsck.gfs2 check the actual bits in the rgrp
bitmaps to see how many blocks are free.  It verifies that
value against the rg_free value in the rgrp.  It needs to
ensure that it starts with an accurate value for rg_free because
an inaccurate value can get us into trouble and cause fsck.gfs2
to abort abnormally.  This can happen in cases where it tries to
allocate blocks (for example, for lost+found metadata) when the
bitmaps say there is free space when there really isn't.

rhbz#455300