libgfs2: Consistent naming for blockmap functions
This patch provides for better consistency in naming of the
functions that operate on the blockmap in memory.  Before,
there were functions like gfs2_bitmap_set and gfs2_set_bitmap
but one operated on the rgrp bitmap and the other operated
on the blockmap in memory.  This patch makes the code more
readable and understandable. It also paves the way for later
simplification of blockmap querying and separating out the
duplicate reference management for performance purposes.

rhbz#455300