libgfs2, fsck.gfs2: simplify block_query code
This patch simplifies the block query code.  Now that the blockmap
code doesn't have the duplicate block reference responsibilities,
we don't need the query info to be in its own structure.  The
block query can be reduced to a simple fetch from memory and mask
off the bits we need.  The block query used to check for a valid
block range.  Now, for performance reasons, that's the
responsibility of the caller.  Most of the callers check for a
valid range anyway, so checking it in the query was a waste of
time.  This is for better performance and makes the code simpler
and more readable.

rhbz#455300