fsck.gfs2: pass1 should use gfs2_special_add not _set
Since pass1 traverses the file system sequentially by block,
it can safely assume blocks it adds to the special_blocks list
are not already on it.  Therefore, it can safely add those
blocks to the list and avoid checking for it already being
there.  This saves a tiny bit of time, which may seem
trivial but when there are a million inodes on the list, the
time saved is worth it.

rhbz#455300