fsck.gfs2: fix directories that have odd number of pointers.
Directory dinodes in gfs2 use the "depth" field to determine the
number of pointers in the hash table.  The depth must be a
factor of 2.  So 2, 4, 8, 16, etc.  That's because the hash table
is doubled every time it gets full.  The problem is, fsck.gfs2
previously had no checks for the depth being correct and/or
the hash table pointers being multiples of two.  Needless to say,
the directory code gets very confused if this condition is not
met due to corruption.  This patch introduces new code to
validate the depth is correct and fix the depth and/or pointers
if they're not.

rhbz#455300