fsck.gfs2: enforce consistency between bitmap and blockmap
One of the big problems with fsck.gfs2 was that there was nothing
to enforce that the incore bitmap from the rgrps corresponded
one-to-one with the incore blockmaps.  This patch enforces that
by making a central function that sets values into the blockmap
and makes sure the bitmap has the correct corresponding bits.
It also tracks all block references with debug messages so block
map problems may be easily tracked down by running fsck.gfs2 with
-vv.  I needed to add some support functions and modified how
the metadata buffers are managed.

rhbz#455300